requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
assemblyai==0.35.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.22.0
ollama==0.4.4
openai-whisper==20240930
pydantic==2.10.3
pydantic_core==2.27.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.3
tiktoken==0.8.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
Werkzeug==3.1.3
zipp==3.21.0
